requirements.txt
alembic==0.9.3.dev0
apturl==0.5.2
asn1crypto==0.24.0
Babel==2.4.0
blinker==1.4
Brlapi==0.6.6
certifi==2019.9.11
chardet==3.0.4
click==6.7
colorama==0.3.7
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
dj-database-url==0.5.0
Django==2.2.7
djangorestframework==3.10.3
Flask==0.12.2
Flask-BabelEx==0.9.3
Flask-Gravatar==0.4.2
Flask-HTMLmin==1.3.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security==1.7.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
gunicorn==20.0.4
htmlmin==0.1.12
httpie==1.0.3
httplib2==0.9.2
idna==2.8
itsdangerous==0.24
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
paramiko==2.0.0
passlib==1.7.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
psutil==5.4.2
psycopg2==2.8.4
pyasn1==0.4.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.4.2
pygobject==3.26.1
pyinotify==0.9.6
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyRFC3339==1.0
python-apt==1.6.4
python-dateutil==2.6.1
python-debian==0.1.32
python-decouple==3.3
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.22.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
SQLAlchemy==1.1.11
sqlparse==0.2.4
sshtunnel==0.1.4
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.7
usb-creator==0.3.3
virtualenv==16.7.8
wadllib==1.3.2
Werkzeug==0.14.1
whitenoise==5.0.1
WTForms==2.1
xkit==0.0.0
zope.interface==4.3.2

